add new card with zoom and draw surface and delete test
