--- faverton-nuxt3/app/pages/test/index.client.vue
@@ -1,7 +1,0 @@
-<template>
-  <div>
-    <ClientOnly>
-      <MapLeaflet />
-    </ClientOnly>
-  </div>
-</template>
